When a scheduled feed update fails, attempt to perform autodiscovery on the website URL.
This is done only if the update failure raises an error that could mean that the URL is unavailable (timeout, host not found, etc). Other kind of errors that could arise e.g. from bugs in the app or problems in the server do not result in an autodiscovery attempt.

If a different fetch_url is discovered from the website, the fetch_url attribute of the feed is updated with the new value and a fetch is attempted from this url, which could result in new entries and in an updated feed title.

This is intended for the case in which a feed owner decides to change its URL, e.g. because he migrates from an in-house solution to feedburner. In these cases usually the website URL does not change, and often the site owner does not even give a warning that the old feed URL is no longer valid. Hopefully he will remember to update the autodiscovery information in his website with the new feed URL.

The old behavior was: the scheduled feed updates started failing, and after 1 week the feed would be marked as permenently unavailable. The feedbunch admin had this week to become aware of the problem and fix it manually by going to the website, copying the new feed URL and updating the feed with it using ActiveAdmin. It was a less than ideal solution that required too much manual intervention.

With the new behavior, feedbunch will try to autodetect the new feed url when this happens. As long as the feed owner updates the autodiscovery information in his website in the week following the URL change, feedbunch will update the feed data and mark the feed as "not failing" automatically, without manual intervention.

Of course if the site owner does not update the autodiscovery data or does not even support autodiscovery, this will fail and manual intervention will still be required. I'd like to send people who do this on a one-way trip to a special circle of hell without internet access, but alas, I have to suffer living in an imperfect world.

+++ lib/feed_autodiscovery.rb
@@ -5,13 +5,14 @@ class FeedAutodiscovery
   ##
   # Try to perform feed autodiscovery on an HTTP response, with the assumption that it's an HTML document.
   #
-  # If successful, save the new feed in the database and return it.
+  # If successful, save the discovered fetch_url in the database and return the updated feed.
   #
   # This method just updates the fetch_url of the feed with the one autodiscovered from the HTML, it doesn't
   # retrieve entries nor do any other changes. It's the responsability of the invoking code to fetch the feed
   # afterwards, populate entries, title, URL etc.
   #
-  # Receives as arguments the feed object to be associated with the discovered feed, and the response with the HTML document.
+  # Receives as arguments the feed object to be associated with the discovered fetch_url, and the response object
+  # with the HTML document.
   #
   # Any errors raised are bubbled to be handled higher up the call chain. In particular, if the response on which
   # autodiscovery is being performed is not an HTML document, an error will be raised.
@@ -42,8 +43,14 @@ class FeedAutodiscovery
 
       # Check if the autodiscovered feed is already in the database
       existing_feed = Feed.url_variants_feed feed_href
-      if existing_feed.present?
-        # If autodiscovered feed is in the database, use it and delete the one passed
+      if existing_feed.present? && existing_feed == feed
+        # The discovered URL is the one the passed feed already has. No changes in the db are necessary.
+        Rails.logger.info "Autodiscovered feed with URL #{feed_href}. Feed #{feed.id} already has this fetch_url, no changes necessary."
+        discovered_feed = feed
+      elsif existing_feed.present? && existing_feed != feed
+        # There is already a feed in the db with the discovered url. Discard the passed feed and return the
+        # already existing feed, so that users are subscribed to the already existing feed instead of having
+        # a duplicate in the db.
         Rails.logger.info "Autodiscovered already known feed with url #{feed_href}. Using it and destroying feed with url #{feed.url} passed as argument"
         feed.destroy
         discovered_feed = existing_feed
